Whew that's a lot.
* Swapped a lot of int primary keys for long, mostly for uniformity partially for paranoia.
* Added the Group model, not wired into the form or repository yet
* Views now have a more abstract ActivityOwner concept which could be a client or a group.
* That ActivityOwner concept exposed a type bug in ActivityEditor - it was getting summary clients not full on clients.

As a note, I originally tried to do this with a ClientActivity and GroupActivity table with all the common fields in a base Activity class. That way both would have a non-null foreign key and avoid the ambiguous cascade delete problem. EF Core 6 HATES this and kept insisting on TPH modeling. EF Core 7 introduces TPC but I don't want to leave the LTS version. With only two owners (group / client) this is probably going to be fine.

+++ TimeTrack/Server/Models/Client.cs
@@ -9,19 +9,19 @@ namespace TimeTrack.Server.Models
      */
     public class Client
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         [MaxLength(16)]
         public string Abbreviation { get; set; }
 
-        public int AgeId { get; set; }
+        public long AgeId { get; set; }
         public Category? Age { get; set; }
-        public int SettingId { get; set; }
+        public long SettingId { get; set; }
         public Category? Setting { get; set; }
-        public int SexualOrientationId { get; set; }
+        public long SexualOrientationId { get; set; }
         public Category? SexualOrientation { get; set; }
-        public int RaceId { get; set; }
+        public long RaceId { get; set; }
         public Category? Race { get; set; }
-        public int GenderId { get; set; }
+        public long GenderId { get; set; }
         public Category? Gender { get; set; }
         public List<Category>? Disabilities { get; set; }
